split e2e file report.tsx
